requirements.txt
openai-whisper==20231117
sounddevice==0.4.6
anyio==4.2.0
discord.py[voice]==2.3.2
keyboard==0.13.5
mutagen==1.46.0
obs_websocket_py==1.0
openai==1.7.2
pydantic==1.10.13
pygame-ce==2.4.0
rich==13.7.0
soundfile==0.12.1
numpy>=1.21.0
